engine/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
smolagents[openai]==1.14.0
pyyaml==6.0.2
sse-starlette==2.3.3
